+++ src/react.ts
@@ -1,0 +1,24 @@
+import * as icons from "./data/icons/react";
+import type { CryptoCurrency, TokenCurrency } from "./types";
+type Icon = React.ComponentType<{
+  size: number;
+  color?: string;
+}>;
+
+function getIconId({ ticker }: CryptoCurrency | TokenCurrency) {
+  let id = ticker.toLowerCase();
+  if (!isNaN(parseInt(id.charAt(0), 10))) id = `_${id}`; // fix variable name leading with a numerical value
+
+  return id;
+}
+
+export function getCryptoCurrencyIcon(
+  currency: CryptoCurrency
+): Icon | null | undefined {
+  return icons[getIconId(currency)];
+}
+export function getTokenCurrencyIcon(
+  token: TokenCurrency
+): Icon | null | undefined {
+  return token.disableCountervalue ? null : icons[getIconId(token)];
+}
